split out components
