+++ client/components/Nav/index.tsx
@@ -9,11 +9,14 @@ const Nav: React.FC = () => {
       fontSize="md"
       width="100%"
       height="60px"
-      justifyContent="flex-start"
+      justifyContent="center"
       alignItems="center"
       boxShadow="5px -8px 15px 5px rgba(0,0,0,0.22)"
+      background="#151515"
     >
-      <Box ml="8">Lystly</Box>
+      <Box color="#fff" ml="8">
+        Lystly
+      </Box>
     </Flex>
   )
 }
