+++ client/package.json
@@ -16,6 +16,7 @@
     "@emotion/styled": "^11.0.0",
     "add": "^2.0.6",
     "axios": "^0.21.0",
+    "firebase": "^8.1.1",
     "formik": "^2.2.5",
     "framer-motion": "^2.9.4",
     "next": "10.0.1",
