--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "firebase": "^8.0.2"
-  }
-}
