+++ server/api_project/api/urls.py
@@ -3,9 +3,9 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('api/v1/user/<int:pk>', views.get_user, name='get_user'),
+    path('api/v1/user/<int:pk>/', views.get_user, name='get_user'),
     path(
-        'api/v1/board/<int:pk>',
+        'api/v1/board/<int:pk>/',
         views.get_delete_update_board,
         name='get_delete_update_board'
     ),
@@ -16,13 +16,13 @@ urlpatterns = [
         name='get_boards'
     ),
     path(
-        'api/v1/card-list/<int:pk>',
+        'api/v1/card-list/<int:pk>/',
         views.get_delete_update_card_list,
         name='get_delete_update_card_list'
     ),
     path('api/v1/card-list/', views.create_card_list, name='create_card_list'),
     path(
-        'api/v1/card/<int:pk>',
+        'api/v1/card/<int:pk>/',
         views.get_delete_update_card,
         name='get_delete_update_card'
     ),
